separate mixins and views
