+++ emgapi/mixins.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from django.shortcuts import get_object_or_404
+
+
+class MultipleFieldLookupMixin(object):
+
+    """
+    Apply this mixin to any view or viewset to get multiple field filtering
+    based on a `lookup_fields` attribute, instead of the default single field
+    filtering.
+    Source: http://www.django-rest-framework.org/api-guide/generic-views/
+    """
+
+    def get_object(self):
+        queryset = self.get_queryset()
+        queryset = self.filter_queryset(queryset)
+        filter = {}
+        for field in self.lookup_fields:
+            if self.kwargs[field]:
+                filter[field] = self.kwargs[field]
+        return get_object_or_404(queryset, **filter)
